requirements.txt
accelerate
fastapi
pillow
torch
torchvision
transformers
uvicorn
numpy==1.26.4